# Improve performance2: reduce unneeded allocation