--- a/go/vimlfunc_test.go
+++ b/go/vimlfunc_test.go
@@ -5,34 +5,6 @@ import (
 	"testing"
 )
 
-func TestViml_eqreg(t *testing.T) {
-	tests := []struct {
-		in   string
-		reg  string
-		want bool
-	}{
-		{in: ``, reg: "^\\s*\\\\", want: false},
-		{in: `hoge`, reg: "^\\s*\\\\", want: false},
-		{in: ` \ hoge`, reg: "^\\s*\\\\", want: true},
-		{in: `\`, reg: "^\\s*\\\\", want: true},
-
-		// ^++
-		{in: `++hoge`, reg: "^++", want: true},
-		{in: `hoge`, reg: "^++", want: false},
-
-		// case
-		{in: `deletel`, reg: "\\v^d%[elete][lp]$", want: true},
-		{in: `deleteL`, reg: "\\v^d%[elete][lp]$", want: true},
-		{in: `++bad=keep`, reg: "^++bad=keep", want: true},
-		{in: `++bad=KEEP`, reg: "^++bad=keep", want: true},
-	}
-	for _, tt := range tests {
-		if got := viml_eqreg(tt.in, tt.reg); got != tt.want {
-			t.Errorf("viml_eqreg(%q, %q) = %v, want %v", tt.in, tt.reg, got, tt.want)
-		}
-	}
-}
-
 func TestViml_eqregh(t *testing.T) {
 	tests := []struct {
 		in   string
@@ -47,24 +19,6 @@ func TestViml_eqregh(t *testing.T) {
 	for _, tt := range tests {
 		if got := viml_eqregh(tt.in, tt.reg); got != tt.want {
 			t.Errorf("viml_eqregh(%q, %q) = %v, want %v", tt.in, tt.reg, got, tt.want)
-		}
-	}
-}
-
-func TestViml_eqregq(t *testing.T) {
-	tests := []struct {
-		in   string
-		reg  string
-		want bool
-	}{
-		{in: `deletel`, reg: "\\v^d%[elete][lp]$", want: true},
-		{in: `deleteL`, reg: "\\v^d%[elete][lp]$", want: true},
-		{in: `++bad=keep`, reg: "^++bad=keep", want: true},
-		{in: `++bad=KEEP`, reg: "^++bad=keep", want: true},
-	}
-	for _, tt := range tests {
-		if got := viml_eqregq(tt.in, tt.reg); got != tt.want {
-			t.Errorf("viml_eqregq(%q, %q) = %v, want %v", tt.in, tt.reg, got, tt.want)
 		}
 	}
 }