ref: https://github.com/haya14busa/go-vimlparser/pull/4