@haya14busa 2016-09-19
### Benchmarks
ref: https://github.com/haya14busa/go-vimlparser/pull/4#issuecomment-247056597

```sh
$ pwd
/home/haya14busa/src/github.com/ynkdir/vim-vimlparser

$ git rev-parse HEAD
2fff43c58968a18bc01bc8304df68bde01af04d9

$ wc -l < autoload/vimlparser.vim
5073

$ go get github.com/haya14busa/go-vimlparser/cmd/vimlparser

$ time vimlparser autoload/vimlparser.vim > /dev/null
vimlparser autoload/vimlparser.vim > /dev/null  0.25s user 0.03s system 114% cpu 0.244 total
```

(autoload/vimlparser.vim is the same as https://github.com/haya14busa/go-vimlparser/pull/4#issuecomment-247056597)

- Before: 0.36s
- After: 0.25s